+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## Unreleased
 
+- added: Support for zero-sync (aka silent-sync) enabled by the change-server protocol.
+- changed: Batch change-server subscriptions for better reliability and efficiency.
 - fixed: Allow pin change while device is in duress mode and user is logged into a non-duress account.
+- fixed: Change conditions for syncNetwork pixies to avoid unnecessary sync and support for zero-sync (aka silent-sync) enabled by the change-server protocol.
 
 ## 2.30.2 (2025-05-30)